feat(通道): MIDI 通道统一承载乐器，XSynth 变为内置可换乐器
- TrackKind 删除 Instrument：轨道只剩 MIDI / 音频两类，
  TrackData 删除 instrument_channel 字段（乐器 Inst 命名空间移除）
- 乐器挂载统一到 256 个 MIDI 通道（A01..P16，global_channel）：
  每通道默认内置 XSynth，可在同一选择器里挂载/切换 CLAP/VST3 插件
- 引擎：ChannelLayout 收敛为 MIDI + 音频两段；插件乐器安装到 MIDI 通道
  dense 并覆盖该通道的 XSynth 输出，卸载即回到 XSynth
- InsertTarget 删除 Instrument 变体；MixerParams.instruments 改为按 MIDI
  通道索引（固定 256），删除乐器通道 strip/inserts/sends 表
- 自动化 PluginParam 的 instrument_channel 改为 MIDI 通道 channel
- MIX：删除独立乐器条，MIDI 通道条内嵌乐器设备行
  （XSynth/插件名 + 界面/参数/更换入口）；dock 同步简化
- 新建轨道对话框删除乐器卡片；mapping/mixer 持久化版本适配（v5）

+++ crates/yinhe-audio/src/audio_renderer.rs
@@ -98,7 +98,7 @@ struct AudioRenderer {
     /// 被替换/移除的 insert 处理器退回 UI 线程（渲染线程不做 deactivate）。
     insert_return_tx: Sender<Vec<Box<dyn yinhe_mixer::InsertProcessor>>>,
     /// 被替换/移除的乐器处理器退回 UI 线程（渲染线程不做 deactivate）。
-    instrument_return_tx: Sender<(u16, Box<dyn yinhe_mixer::InstrumentProcessor>)>,
+    instrument_return_tx: Sender<(u8, Box<dyn yinhe_mixer::InstrumentProcessor>)>,
     /// 是否启用 GPU 合成器。启用后加载音色库时初始化 GpuSynth，渲染走 engine.gpu_synth。
     #[cfg(feature = "gpu")]
     use_gpu_synth: bool,
@@ -129,7 +129,7 @@ impl AudioRenderer {
         // cpal 回调每次请求的帧数（预览时 ring 目标下限，避免回调欠载静音）。
         callback_frames: usize,
         insert_return_tx: Sender<Vec<Box<dyn yinhe_mixer::InsertProcessor>>>,
-        instrument_return_tx: Sender<(u16, Box<dyn yinhe_mixer::InstrumentProcessor>)>,
+        instrument_return_tx: Sender<(u8, Box<dyn yinhe_mixer::InstrumentProcessor>)>,
         #[cfg(feature = "gpu")] use_gpu_synth: bool,
     ) -> Self {
         Self {
@@ -1108,7 +1108,7 @@ pub(crate) fn spawn_renderer(
     // cpal 回调每次请求的帧数（预览时 ring 目标下限）。
     callback_frames: usize,
     insert_return_tx: Sender<Vec<Box<dyn yinhe_mixer::InsertProcessor>>>,
-    instrument_return_tx: Sender<(u16, Box<dyn yinhe_mixer::InstrumentProcessor>)>,
+    instrument_return_tx: Sender<(u8, Box<dyn yinhe_mixer::InstrumentProcessor>)>,
     #[cfg(feature = "gpu")] use_gpu_synth: bool,
 ) -> Result<JoinHandle<()>, std::io::Error> {
     thread::Builder::new()
@@ -1151,7 +1151,7 @@ pub(crate) fn spawn_renderer(
                 let _ = insert_return_tx.send(leftovers);
             }
             // 引擎里仍在位的乐器处理器也退回 UI 线程回收。
-            let inst_leftovers: Vec<(u16, Box<dyn yinhe_mixer::InstrumentProcessor>)> = renderer
+            let inst_leftovers: Vec<(u8, Box<dyn yinhe_mixer::InstrumentProcessor>)> = renderer
                 .engine
                 .instruments
                 .iter_mut()
